pyproject.toml
[tool.poetry]
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.121.2"
uvicorn = {extras = ["standard"], version = "^0.38.0"}
pydantic = {extras = ["email"], version = "^2.12.4"}
pydantic-settings = "^2.12.0"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.44"}
asyncpg = "^0.30.0"
alembic = "^1.17.2"
python-jose = {extras = ["cryptography"], version = "^3.5.0"}
argon2-cffi = "^25.1.0"
passlib = {extras = ["argon2"], version = "^1.7.4"}

[dependency-groups]
dev = [
    "black (>=25.11.0,<26.0.0)"
]
